+++ perl/FSM/SourceClassifier.pm
@@ -13,6 +13,8 @@ sub classify_source_ast ($raw_ast) {
     if (defined $header) {
         if ($header eq '+fsm' || $header =~ /^\?fsm:/) {
             $kind = 'fsm';
+        } elsif ($header =~ /^\?dt:/) {
+            $kind = 'dt';
         } elsif ($header =~ /^\?top:/) {
             $kind = 'composition';
         }